Extract common parts of autocomplete manager and autocomplete provider
