+++ src/main/kotlin/com/sourcegraph/cody/autocomplete/CodyAutocompleteManager.kt
@@ -15,7 +15,6 @@ import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.editor.Inlay
 import com.intellij.openapi.editor.InlayModel
 import com.intellij.openapi.editor.colors.EditorColorsManager
-import com.intellij.openapi.fileEditor.FileDocumentManager
 import com.intellij.openapi.keymap.KeymapUtil
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.popup.Balloon
@@ -28,29 +27,15 @@ import com.intellij.util.concurrency.annotations.RequiresEdt
 import com.sourcegraph.cody.Icons
 import com.sourcegraph.cody.agent.CodyAgentService
 import com.sourcegraph.cody.agent.protocol.AutocompleteItem
-import com.sourcegraph.cody.agent.protocol.AutocompleteParams
 import com.sourcegraph.cody.agent.protocol.AutocompleteResult
-import com.sourcegraph.cody.agent.protocol.AutocompleteTriggerKind
 import com.sourcegraph.cody.agent.protocol.CompletionItemParams
-import com.sourcegraph.cody.agent.protocol.ErrorCode
-import com.sourcegraph.cody.agent.protocol.ErrorCodeUtils.toErrorCode
-import com.sourcegraph.cody.agent.protocol.ProtocolTextDocument.Companion.uriFor
-import com.sourcegraph.cody.agent.protocol.RateLimitError.Companion.toRateLimitError
-import com.sourcegraph.cody.agent.protocol.SelectedCompletionInfo
-import com.sourcegraph.cody.agent.protocol_extensions.Position
-import com.sourcegraph.cody.agent.protocol_generated.Position
-import com.sourcegraph.cody.agent.protocol_generated.Range
+import com.sourcegraph.cody.autocomplete.Utils.triggerAutocompleteAsync
 import com.sourcegraph.cody.autocomplete.render.AutocompleteRendererType
 import com.sourcegraph.cody.autocomplete.render.CodyAutocompleteBlockElementRenderer
 import com.sourcegraph.cody.autocomplete.render.CodyAutocompleteElementRenderer
 import com.sourcegraph.cody.autocomplete.render.CodyAutocompleteSingleLineRenderer
 import com.sourcegraph.cody.autocomplete.render.InlayModelUtil.getAllInlaysForEditor
 import com.sourcegraph.cody.config.CodyAuthenticationManager
-import com.sourcegraph.cody.ignore.ActionInIgnoredFileNotification
-import com.sourcegraph.cody.ignore.IgnoreOracle
-import com.sourcegraph.cody.ignore.IgnorePolicy
-import com.sourcegraph.cody.statusbar.CodyStatus
-import com.sourcegraph.cody.statusbar.CodyStatusService.Companion.notifyApplication
 import com.sourcegraph.cody.statusbar.CodyStatusService.Companion.resetApplication
 import com.sourcegraph.cody.vscode.CancellationToken
 import com.sourcegraph.cody.vscode.InlineCompletionTriggerKind
@@ -58,7 +43,6 @@ import com.sourcegraph.cody.vscode.IntelliJTextDocument
 import com.sourcegraph.cody.vscode.TextDocument
 import com.sourcegraph.common.CodyBundle
 import com.sourcegraph.common.CodyBundle.fmt
-import com.sourcegraph.common.UpgradeToCodyProNotification
 import com.sourcegraph.config.ConfigUtil.isCodyEnabled
 import com.sourcegraph.config.UserLevelConfig
 import com.sourcegraph.utils.CodyEditorUtil.getAllOpenEditors
@@ -68,13 +52,8 @@ import com.sourcegraph.utils.CodyEditorUtil.isCommandExcluded
 import com.sourcegraph.utils.CodyEditorUtil.isEditorValidForAutocomplete
 import com.sourcegraph.utils.CodyEditorUtil.isImplicitAutocompleteEnabledForEditor
 import com.sourcegraph.utils.CodyFormatter
-import java.util.concurrent.CancellationException
-import java.util.concurrent.CompletableFuture
-import java.util.concurrent.CompletionException
-import java.util.concurrent.TimeUnit
 import java.util.concurrent.atomic.AtomicReference
 import java.util.stream.Collectors
-import org.eclipse.lsp4j.jsonrpc.ResponseErrorException
 
 /** Responsible for triggering and clearing inline code completions (the autocomplete feature). */
 @Service
@@ -199,110 +178,11 @@ class CodyAutocompleteManager {
         triggerKind,
         cancellationToken,
         lookupString,
-        originalText)
-  }
-
-  /** Asynchronously triggers auto-complete for the given editor and offset. */
-  private fun triggerAutocompleteAsync(
-      project: Project,
-      editor: Editor,
-      offset: Int,
-      textDocument: TextDocument,
-      triggerKind: InlineCompletionTriggerKind,
-      cancellationToken: CancellationToken,
-      lookupString: String?,
-      originalText: String
-  ): CompletableFuture<Void?> {
-    val position = textDocument.positionAt(offset)
-    val lineNumber = editor.document.getLineNumber(offset)
-    var startPosition = 0
-    if (!lookupString.isNullOrEmpty()) {
-      startPosition = findLastCommonSuffixElementPosition(originalText, lookupString)
-    }
-
-    val virtualFile =
-        FileDocumentManager.getInstance().getFile(editor.document)
-            ?: return CompletableFuture.completedFuture(null)
-    val params =
-        if (lookupString.isNullOrEmpty())
-            AutocompleteParams(
-                uriFor(virtualFile),
-                Position(position.line, position.character),
-                if (triggerKind == InlineCompletionTriggerKind.INVOKE)
-                    AutocompleteTriggerKind.INVOKE.value
-                else AutocompleteTriggerKind.AUTOMATIC.value)
-        else
-            AutocompleteParams(
-                uriFor(virtualFile),
-                Position(position.line, position.character),
-                AutocompleteTriggerKind.AUTOMATIC.value,
-                SelectedCompletionInfo(
-                    lookupString,
-                    if (startPosition < 0) Range(position, position)
-                    else Range(Position(lineNumber, startPosition), position)))
-    notifyApplication(project, CodyStatus.AutocompleteInProgress)
-
-    val resultOuter = CompletableFuture<Void?>()
-    CodyAgentService.withAgent(project) { agent ->
-      if (triggerKind == InlineCompletionTriggerKind.INVOKE &&
-          IgnoreOracle.getInstance(project).policyForUri(virtualFile.url, agent).get() !=
-              IgnorePolicy.USE) {
-        ActionInIgnoredFileNotification.maybeNotify(project)
-        resetApplication(project)
-        resultOuter.cancel(true)
-      } else {
-        val completions = agent.server.autocompleteExecute(params)
-
-        // Important: we have to `.cancel()` the original `CompletableFuture<T>` from lsp4j. As soon
-        // as we use `thenAccept()` we get a new instance of `CompletableFuture<Void>` which does
-        // not
-        // correctly propagate the cancellation to the agent.
-        cancellationToken.onCancellationRequested { completions.cancel(true) }
-
-        ApplicationManager.getApplication().executeOnPooledThread {
-          completions
-              .handle { result, error ->
-                if (error != null) {
-                  if (triggerKind == InlineCompletionTriggerKind.INVOKE ||
-                      !UpgradeToCodyProNotification.isFirstRLEOnAutomaticAutocompletionsShown) {
-                    handleError(project, error)
-                  }
-                } else if (result != null && result.items.isNotEmpty()) {
-                  UpgradeToCodyProNotification.isFirstRLEOnAutomaticAutocompletionsShown = false
-                  UpgradeToCodyProNotification.autocompleteRateLimitError.set(null)
-                  processAutocompleteResult(editor, offset, triggerKind, result, cancellationToken)
-                }
-                null
-              }
-              .exceptionally { error: Throwable? ->
-                if (!(error is CancellationException || error is CompletionException)) {
-                  logger.warn("failed autocomplete request $params", error)
-                }
-                null
-              }
-              .completeOnTimeout(null, 3, TimeUnit.SECONDS)
-              .thenRun { // This is a terminal operation, so we needn't call get().
-                resetApplication(project)
-                resultOuter.complete(null)
-              }
+        originalText,
+        logger) { autocompleteResult ->
+          processAutocompleteResult(
+              editor, offset, triggerKind, autocompleteResult, cancellationToken)
         }
-      }
-    }
-    cancellationToken.onCancellationRequested { resultOuter.cancel(true) }
-    return resultOuter
-  }
-
-  private fun handleError(project: Project, error: Throwable?) {
-    if (error is ResponseErrorException) {
-      if (error.toErrorCode() == ErrorCode.RateLimitError) {
-        val rateLimitError = error.toRateLimitError()
-        UpgradeToCodyProNotification.autocompleteRateLimitError.set(rateLimitError)
-        UpgradeToCodyProNotification.isFirstRLEOnAutomaticAutocompletionsShown = true
-        ApplicationManager.getApplication().executeOnPooledThread {
-          UpgradeToCodyProNotification.notify(error.toRateLimitError(), project)
-        }
-      }
-    }
   }
 
   private fun processAutocompleteResult(
@@ -511,21 +391,6 @@ class CodyAutocompleteManager {
     val lineSpacing = fontPreferences.lineSpacing.toInt()
     val extraMargin = 4
     return fontSize + lineSpacing + extraMargin
-  }
-
-  private fun findLastCommonSuffixElementPosition(
-      stringToFindSuffixIn: String,
-      suffix: String
-  ): Int {
-    var i = 0
-    while (i <= suffix.length) {
-      val partY = suffix.substring(0, suffix.length - i)
-      if (stringToFindSuffixIn.endsWith(partY)) {
-        return stringToFindSuffixIn.length - (suffix.length - i)
-      }
-      i++
-    }
-    return 0
   }
 
   private fun cancelCurrentJob(project: Project?) {
